+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ Unreleased
 Added
 -----
 
+- **Irodsbackend**
+    - Create iRODS user accounts at login for users with LDAP/SODAR auth (#1315)
 - **Landingzones**
     - Optional zone write access restriction to created collections (#1050)
 - **Samplesheets**